Rename HeatSeek.h to game-source-code/HeatSeek.h
